# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ledger/opening-balances": {
      "post": {
        "operationId": "OpeningBalancesController_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpeningBalancesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Journal"
        ]
      }
    },
...
    "/sales-invoices": {
      "get": {
        "operationId": "SalesInvoicesController_list",
        "parameters": [
          {
            "name": "partnerId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "POSTED",
                "VOID"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Sales Invoices"
        ]
      },
      "post": {
        "operationId": "SalesInvoicesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSalesInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Sales Invoices"
        ]
      }
    },
    "/sales-invoices/{id}": {
      "get": {
        "operationId": "SalesInvoicesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Sales Invoices"
        ]
      },
      "patch": {
        "operationId": "SalesInvoicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSalesInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Sales Invoices"
        ]
      },
      "delete": {
        "operationId": "SalesInvoicesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Sales Invoices"
        ]
      }
    },
    "/sales-invoices/{id}/post": {
      "post": {
        "operationId": "SalesInvoicesController_post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Sales Invoices"
        ]
      }
    },
    "/sales-invoices/{id}/void": {
      "post": {
        "operationId": "SalesInvoicesController_void",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Sales Invoices"
        ]
      }
    },
...
    "/payments": {
      "get": {
        "operationId": "PaymentsController_list",
        "parameters": [
          {
            "name": "partnerId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "RECEIPT",
                "DISBURSEMENT"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "POSTED",
                "VOID"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      },
      "post": {
        "operationId": "PaymentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      }
    },
    "/payments/{id}": {
      "get": {
        "operationId": "PaymentsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      },
      "delete": {
        "operationId": "PaymentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      }
    },
    "/payments/{id}/post": {
      "post": {
        "operationId": "PaymentsController_post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      }
    },
    "/payments/{id}/void": {
      "post": {
        "operationId": "PaymentsController_void",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Payments"
        ]
      }
    },
...
    "/close/year-end": {
      "post": {
        "operationId": "ClosingController_run",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseYearDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Close"
        ]
      }
    },
    "/close/year-end/{fiscalYear}/reopen": {
      "post": {
        "operationId": "ClosingController_reopen",
        "parameters": [
          {
            "name": "fiscalYear",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Close"
        ]
      }
    },
    "/close/year-end/{fiscalYear}": {
      "get": {
        "operationId": "ClosingController_status",
        "parameters": [
          {
            "name": "fiscalYear",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Close"
        ]
      }
    },
...
  },
  "info": {
    "title": "Indonesian Accounting API",
    "description": "Conventions, roles, and lifecycles: see docs/api/frontend-guide.md",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "JournalLineDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "debit": {
            "type": "string"
          },
          "credit": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "accountId"
        ]
      },
...
      "OpeningBalancesDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "balances": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JournalLineDto"
            }
          }
        },
        "required": [
          "date",
          "balances"
        ]
      },
...
      "SalesInvoiceLineDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 255
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d{1,4})?$"
          },
          "unitPrice": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d{1,4})?$"
          },
          "taxCodeIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "description",
          "accountId",
          "quantity",
          "unitPrice",
          "taxCodeIds"
        ]
      },
      "CreateSalesInvoiceDto": {
        "type": "object",
        "properties": {
          "partnerId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "maxLength": 255
          },
          "lines": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SalesInvoiceLineDto"
            }
          }
        },
        "required": [
          "partnerId",
          "date",
          "lines"
        ]
      },
      "UpdateSalesInvoiceDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "maxLength": 255
          },
          "lines": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SalesInvoiceLineDto"
            }
          }
        }
      },
...
      "AllocationDto": {
        "type": "object",
        "properties": {
          "salesInvoiceId": {
            "type": "string",
            "format": "uuid"
          },
          "purchaseBillId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d{1,4})?$"
          }
        },
        "required": [
          "amount"
        ]
      },
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "direction": {
            "type": "string",
            "enum": [
              "RECEIPT",
              "DISBURSEMENT"
            ]
          },
          "partnerId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string"
          },
          "cashAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "description": {
            "type": "string",
            "maxLength": 255
          },
          "allocations": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AllocationDto"
            }
          }
        },
        "required": [
          "direction",
          "partnerId",
          "date",
          "cashAccountId",
          "allocations"
        ]
      },
      "CloseYearDto": {
        "type": "object",
        "properties": {
          "fiscalYear": {
            "type": "number",
            "minimum": 2000,
            "maximum": 2100
          }
        },
        "required": [
          "fiscalYear"
        ]
      }
    }
  }
}